4Men1Joke/API/requirements.txt
asgiref==3.5.0
bcrypt==3.2.2
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
firebase-admin==5.2.0
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.4
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-secret-manager==2.10.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.3
Pillow==9.1.1
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.8
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
